+++ Assets/Nova/Sources/Core/Collections/ByteSegment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbb972dffca834c46903daf43a31d6ee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
